fix: change Entity primary key to id: Long

+++ src/main/java/io/muxfe/springbootstarter/entity/User.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
@@ -19,10 +18,11 @@ public class User {
   private String password;
   @ElementCollection(fetch = FetchType.EAGER)
   private List<String> roles;
+
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  private Integer id;
-  
+  private Long id;
+
   public User() {
   }
 }
