+++ src/main/java/io/muxfe/springbootstarter/repository/EmployeeRepository.java
@@ -3,6 +3,6 @@ package io.muxfe.springbootstarter.repository;
 import io.muxfe.springbootstarter.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
   
 }
